apps/api/pyproject.toml
[project]
name = "zb-vaka-api"
version = "0.1.0"
description = "ZB Vaka FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.14.1",
  "email-validator==2.2.0",
  "fastapi==0.115.6",
  "passlib[bcrypt]==1.7.4",
  "psycopg[binary]==3.2.3",
  "pydantic-settings==2.7.0",
  "python-jose[cryptography]==3.3.0",
  "redis==5.2.1",
  "sqlalchemy==2.0.36",
  "uvicorn[standard]==0.32.1",
  "pytest==8.3.4",
  "pytest-asyncio==0.25.0",
  "httpx==0.28.1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
